python-selenium-master/requirements.txt
selenium
ddt
pyyaml